requirements.txt
# -------------------------
# Core LangChain packages
# -------------------------
langchain>=0.2.0
langchain-core>=0.2.0
langchain-community>=0.2.0
langchain-openai>=0.1.0
langchain-pinecone>=0.1.0
langchain-text-splitters>=0.2.0

# -------------------------
# Pinecone (latest SDK v3+)
# -------------------------
pinecone-client>=3.0.0

# -------------------------
# Azure + Tokenization
# -------------------------
tiktoken>=0.7.0

# -------------------------
# PDF + Utility packages
# -------------------------
pypdf>=4.0.0
python-dotenv>=1.0.0
tqdm>=4.66.0

# -------------------------
# Web UI
# -------------------------
streamlit>=1.38.0
